requirements.txt
bilibili_api_python==16.2.0
discord.py==2.3.2
python-dotenv==1.0.1
yt_dlp==2024.4.9
pynacl==1.5.0